The  file contains information about your coding activity, including the number of files modified, the lines added or deleted, and the time spent on coding activities. This file is used to generate reports that summarize your coding activity and help you keep track of your progress. The deps-cleaner is a cli that helps you in cleaning your space by removing dependencies folders that can be recreated very easily.
Usage:
  deps-cleaner [flags]
  deps-cleaner [command]

Available Commands:
  clean         Clean dependencies
  completion    Generate the autocompletion script for the specified shell
  help          Help about any command
  update:config Update local configuration with remote configuration

Flags:
  -h, --help      help for deps-cleaner
  -v, --version   Prints the version

Use "deps-cleaner [command] --help" for more information about a command. directory contains a project to clean up unused dependencies in your codebase. The  file in this directory provides information about how to use the tool and its features.
The  directory contains a project for a French language learning platform. The  file in this directory provides information about the project's requirements, technologies used, and how to run it locally.
The  directory contains a collection of solutions to LeetCode problems that you have solved. The  file in this directory provides information about the challenges you have completed and their corresponding solutions.
The  directory contains an extension for Visual Studio Code that tracks your coding activity and allows you to generate reports summarizing your progress. The  file in this directory provides information about how to install and use the extension, as well as its features and capabilities.
The  directory contains a project for a web application using Next.js. The  file in this directory provides information about the project's requirements, technologies used, and how to run it locally.
The  directory contains a simple content management system (CMS) built with React and TinyMCE. The  file in this directory provides information about the project's requirements, technologies used, and how to run it locally.