Here is a concise commit message based on the provided git diff and modified files:
In this message, we are using the  type to indicate that we made changes to the codebase without adding new functionality. The scope of this change is limited to the  project.
The body of the commit message explains the change in more detail. We mention that we updated the active time calculation for 8 minutes, which likely means that we corrected a previous mistake or added some new logic to our codebase.
Finally, we include the
wtmp begins Fri Oct 25 22:32:43 CEST 2024 timestamp at the end of the message, indicating when this commit was made.