The commit message is:
The commit message highlights the "fix" type, which indicates that the modification was intended to fix a bug or problem in the code. It also provides an overview of the changes made on January 3rd by mentioning the files modified and the type of change made. The message concludes with the duration of activity for the day.